pyproject.toml
[tool.poetry]
name = "tgfp"
version = "11.1.0"
description = "Scripts to help with the running of The Great Football Pool"
authors = ["John Sturgeon <john.sturgeon@me.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.11,<4"
tgfp-lib = "^2.2.2"
tgfp-nfl = "^6.1.0"
arrow = "^1.2.3"
sentry-sdk = "^1.9.8"
pylint = "^3.0.2"
prefect = "^2.13.1"
greenlet = "^3.0.0"
httpx = "^0.25.0"
pytest = "^7.4.3"
discord-py = "^2.3.2"
aiomqtt = "^1.2.1"
python-dotenv = "^1.0.0"
rich-cli = "^1.8.0"
matplotlib = "^3.8.2"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
